fix snapshot and data type

+++ shared/postgres/docker-entrypoint-initdb.d/create_table_employees.sql
@@ -7,7 +7,7 @@ CREATE TABLE employees (
     full_time BOOLEAN NOT NULL DEFAULT TRUE,
     position VARCHAR(100),
     hire_date DATE NOT NULL,
-    salary DECIMAL(10, 2) CHECK (salary >= 0),
+    salary REAL CHECK (salary >= 0),
     updated_at TIMESTAMPTZ DEFAULT NOW(),
     created_at TIMESTAMPTZ DEFAULT NOW(),
     PRIMARY KEY (id)
